task context

+++ src/components/Home/Block/Options/OptionsMenu/index.tsx
@@ -1,30 +1,27 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { Button, notification } from 'antd';
-import { TTask } from 'constants/types/common';
 import { useAppDispatch } from 'customHooks/redux/useAppDispatch';
 import { useAppSelector } from 'customHooks/redux/useAppSelector';
 import { canUserDuplicateTask, isUserTaskAuthor } from 'helpers/userRoles';
 import { getVerifyIdUser } from 'store/auth/verify/selectors';
 import { cloneTaskAction } from 'store/createTask/thunk';
 import { deleteTaskAction } from 'store/tasks/thunk';
+import { TaskContext } from 'constants/taskContext';
 import styles from './index.module.scss';
 
-interface IProps {
-  task: TTask;
-}
-
-const OptionsMenu: React.FC<IProps> = ({ task }) => {
+const OptionsMenu: React.FC = () => {
   const dispatch = useAppDispatch();
   const userId = useAppSelector(getVerifyIdUser);
+  const task = useContext(TaskContext);
 
   const handleCloneTask = (): void => {
-    if (canUserDuplicateTask(userId, task)) {
+    if (task && canUserDuplicateTask(userId, task)) {
       dispatch(cloneTaskAction({ id: task.task_id, edit: false }));
     } else notification.warn({ message: 'Нет прав на дублирование задачи' });
   };
 
   const handleDeleteTask = (): void => {
-    if (isUserTaskAuthor(userId, task)) {
+    if (task && isUserTaskAuthor(userId, task)) {
       dispatch(deleteTaskAction(task.task_id));
     } else notification.warn({ message: 'Удалить задачу может только автор' });
   };
